backend/requirements.txt
flask===1.1.1
mongoengine==0.18.2
python-dotenv==0.10.3
Pillow==5.2.0
pymongo==3.5.1
Wand==0.4.4
colorama==0.3.9
tqdm==4.31.1
flask-cors==3.0.8
gunicorn==19.9.0
flask-restful==0.3.7
qrcode==6.0